feat(callback): implement balanced energy calculation using callback pattern
Implemented interface-driven callbacks to handle input addition, result computation, and
reset operations for the balanced energy module. This decouples computation logic from
the UI layer and ensures clear data flow. State persistence across configuration changes
is maintained through onSaveInstanceState.

Renamed package architecting_state to architectingstate for lint compliance.
Updated wildcard imports from test files.
