+++ app/src/main/java/com/karthik/pro/engr/architectingstate/domain/energy/EnergyAnalyzer.kt
@@ -1,7 +1,4 @@
-package com.karthik.pro.engr.lib.domain.energy
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
+package com.karthik.pro.engr.architectingstate.domain.energy
 
 object EnergyAnalyzer {
     fun findLongestStretch(houseTypes: List<String>): StretchResult {
@@ -37,8 +34,9 @@ object EnergyAnalyzer {
                 } result-> ${result.joinToString(", ")}")
         }
 
+//        "The Longest Stretch Houses Starts from ${result[0] + 1} to ${result[1] + 1}"
         return StretchResult(result[0] + 1, result[1], maxLen)
     }
 }
-@Parcelize
-data class StretchResult(val startIndex: Int, val endIndex: Int, val length: Int) : Parcelable
+
+data class StretchResult(val startIndex: Int, val endIndex: Int, val length: Int)
