+++ app/src/test/java/com/karthik/pro/engr/architectingstate/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.karthik.pro.engr.res
+package com.karthik.pro.engr.architectingstate
 
 import org.junit.Test
