+++ callback/src/main/java/com/karthik/pro/engr/callback/EnergyListener.kt
@@ -1,0 +1,11 @@
+package com.karthik.pro.engr.callback
+
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.setValue
+
+interface EnergyListener {
+
+    fun addHouseType(type: String)
+    fun computeResult()
+    fun reset()
+}
