+++ callback/src/test/java/com/karthik/pro/engr/callback/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.karthik.pro.engr.res
+package com.karthik.pro.engr.callback
 
 import org.junit.Test
