+++ settings.gradle.kts
@@ -21,4 +21,5 @@ dependencyResolutionManagement {
 
 rootProject.name = "architecting-state"
 include(":app")
+include(":callback")
 include(":statelib")
